+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Object has different properties!
 - Support for binaries.
 - Support for parsing extra multiparts.
 - Support for History-Info header.
+- Support platforms without support for `globalThis`.
 - Ability to remove message listeners.
 ### Fixed
 - Conversation state only changes if the state value itself changes. If origin is different but state is the same, it won't trigger a state change.